+++ pkg/ted25519/frost/challenge_derive.go
@@ -7,6 +7,7 @@
 package frost
 
 import (
+	"crypto/sha256"
 	"crypto/sha512"
 
 	"github.com/TEENet-io/kryptology/pkg/core/curves"
@@ -30,9 +31,9 @@ type Secp256k1ChallengeDeriver struct{}
 
 // DeriveChallenge implements the FROST challenge derivation for secp256k1 using SHA-256.
 func (d Secp256k1ChallengeDeriver) DeriveChallenge(msg []byte, pubKey curves.Point, r curves.Point) (curves.Scalar, error) {
-	h := sha512.New()
+	h := sha256.New()
 	_, _ = h.Write(r.ToAffineCompressed())
 	_, _ = h.Write(pubKey.ToAffineCompressed())
 	_, _ = h.Write(msg)
-	return new(curves.ScalarK256).SetBytesWide(h.Sum(nil))
+	return new(curves.ScalarK256).SetBytes(h.Sum(nil))
 }